botService actions; move Trash and Polls action;
