+++ internal/misc/misc.go
@@ -1,0 +1,10 @@
+package misc
+
+import (
+	"reflect"
+	"runtime"
+)
+
+func GetFunctionName(i interface{}) string {
+	return runtime.FuncForPC(reflect.ValueOf(i).Pointer()).Name()
+}
